// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ActivityType = {
  demo: 'demo',
  call: 'call',
  email: 'email'
} as const

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType]


export const Segment = {
  SMB: 'SMB',
  MidMarket: 'MidMarket',
  Enterprise: 'Enterprise'
} as const

export type Segment = (typeof Segment)[keyof typeof Segment]


export const DealStage = {
  ClosedWon: 'ClosedWon',
  Prospecting: 'Prospecting',
  Negotiation: 'Negotiation',
  ClosedLost: 'ClosedLost'
} as const

export type DealStage = (typeof DealStage)[keyof typeof DealStage]


export const Industry = {
  SaaS: 'SaaS',
  Ecommerce: 'Ecommerce',
  FinTech: 'FinTech',
  EdTech: 'EdTech',
  Healthcare: 'Healthcare'
} as const

export type Industry = (typeof Industry)[keyof typeof Industry]